backend/requirements.txt
fastapi==0.115.11
uvicorn==0.30.6
sqlalchemy==2.0.36
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic[email]==2.9.2
alembic==1.13.3